requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# OCR and PDF processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# LLM and API
httpx==0.26.0
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3

# Utilities
python-magic-bin==0.4.14; platform_system == "Windows"
